src/backend/requirements.txt
# Web framework
requests==2.32.5
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database + ORM
SQLAlchemy==2.0.35
psycopg2-binary
pydantic[email]

# Migrations
alembic==1.13.2

# Env management
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2

# Authentication
PyJWT==2.10.1
bcrypt==4.3.0

# Location stuff
GeoAlchemy2==0.18.0
shapely==2.1.2

# Riddle stuff
openai==2.1.0